Make setup work better with chains When adding controls, re-parent the BBbone_Start controllers to the bone's parent's BBone_End controller, and hide the start controller. This will make it so you only need to move one controller, instead of two disconnect controllers, when they're supposed to always stick together.
Another thing is when adding Start/End and In/Out controllers, separately.
When they're added together, the In/Out controllers are parented to the Start/End controllers.
If you add them in different runs, they would retain the parenting from adding them separately.
Now, they will be parented as intended in subsequent runs.